chore: merge fedora and upstream spec file
The upstream splits out things that don't really need
to be, but have things the Fedora spec doesn't. Merge
the two for the best of both worlds so we can use things
like packit to automate as much as possible. We also
drop the Python2 builds as it's long EOL.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>